problem : security에서 NullPointerException 핸들 에러

+++ Board/src/main/java/com/sparta/board/service/CommentService.java
@@ -6,16 +6,15 @@ import com.sparta.board.entity.Board;
 import com.sparta.board.entity.Comment;
 import com.sparta.board.entity.User;
 import com.sparta.board.entity.UserRoleEnum;
+import com.sparta.board.exception.CustomException;
+import com.sparta.board.exception.ErrorCode;
 import com.sparta.board.jwt.JwtUtil;
 import com.sparta.board.repository.BoardRepository;
 import com.sparta.board.repository.CommentRepository;
 import com.sparta.board.repository.UserRepository;
-import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import javax.servlet.http.HttpServletRequest;
 
 @Service
 @RequiredArgsConstructor
@@ -27,32 +26,30 @@ public class CommentService {
     private final UserRepository userRepository;
 
     @Transactional
-    public CommentResponseDto createComment(CommentRequestDto requestDto, HttpServletRequest request) {
-        User user = authenticateUser(request);
+    public CommentResponseDto createComment(CommentRequestDto requestDto, User user) {
         Board board = boardRepository.findById(requestDto.getBoardId()).orElseThrow(
-                () -> new NullPointerException("해당 글이 없습니다.")
+                () -> new CustomException(ErrorCode.NULL_BOARD)
         );
         Comment comment = commentRepository.saveAndFlush(new Comment(requestDto, board, user));
         return new CommentResponseDto(comment);
     }
 
     @Transactional
-    public CommentResponseDto updateComment(Long cmtId, CommentRequestDto requestDto, HttpServletRequest request) {
-        User user = authenticateUser(request);
+    public CommentResponseDto updateComment(Long cmtId, CommentRequestDto requestDto, User user) {
         Board board = boardRepository.findById(requestDto.getBoardId()).orElseThrow(
-                () -> new NullPointerException("해당 글이 없습니다.")
+                () -> new CustomException(ErrorCode.NULL_BOARD)
         );
         UserRoleEnum userRoleEnum = user.getRole();
 
         if (userRoleEnum == UserRoleEnum.USER) {
             Comment comment = commentRepository.findByIdAndUserId(cmtId, user.getId()).orElseThrow(
-                    () -> new IllegalArgumentException("해당 댓글이 없거나, 수정할 권한이 없습니다.")
+                    () -> new CustomException(ErrorCode.NULL_BOARD)
             );
             comment.updateComment(requestDto, board);
             return new CommentResponseDto(comment);
         } else {
             Comment comment = commentRepository.findById(cmtId).orElseThrow(
-                    () -> new NullPointerException("해당 댓글이 없습니다.")
+                    () -> new CustomException(ErrorCode.NULL_BOARD)
             );
             comment.updateComment(requestDto, board);
             return new CommentResponseDto(comment);
@@ -60,40 +57,23 @@ public class CommentService {
     }
 
     @Transactional
-    public void deleteComment(Long cmtId, HttpServletRequest request) {
-        User user = authenticateUser(request);
+    public void deleteComment(Long cmtId, User user) {
         UserRoleEnum userRoleEnum = user.getRole();
 
         if (userRoleEnum == UserRoleEnum.USER) {
-            Comment comment = commentRepository.findByIdAndUserId(cmtId, user.getId()).orElseThrow(
-                    () -> new IllegalArgumentException("해당 댓글이 없거나, 삭제할 권한이 없습니다.")
+            Comment comment = commentRepository.findById(cmtId).orElseThrow(
+                    () -> new CustomException(ErrorCode.NULL_BOARD)
             );
+
+            if (!comment.getUser().getId().equals(user.getId())) {
+                throw new CustomException(ErrorCode.INVALID_USER);
+            }
             commentRepository.delete(comment);
         } else {
             Comment comment = commentRepository.findById(cmtId).orElseThrow(
-                    () -> new NullPointerException("해당 댓글이 없습니다.")
+                    () -> new CustomException(ErrorCode.NULL_BOARD)
             );
             commentRepository.delete(comment);
-        }
-    }
-
-
-
-    private User authenticateUser(HttpServletRequest request) {
-        String token = jwtUtil.resolveToken(request);
-        Claims claims;
-
-        if (token != null) {
-            if (jwtUtil.validateToken(token)) {
-                claims = jwtUtil.getUserInfoFromToken(token);
-            } else {
-                throw new IllegalArgumentException("Token Error");
-            }
-            return userRepository.findByUsername(claims.getSubject()).orElseThrow(
-                    () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
-            );
-        } else {
-            throw new IllegalArgumentException("토큰이 없습니다.");
         }
     }
 }
